Refactor - Replace temp with Query, minor changes
