+++ src/main/java/eu/assuremoss/VulnRepairDriver.java
@@ -86,6 +86,8 @@ public class VulnRepairDriver {
         MLOG.info(String.format("Detected %d vulnerabilities", vulnerabilityLocations.size()));
         vulnerabilityLocations.forEach(vulnEntry -> MLOG.fInfo(vulnEntry.getType() + " -> " + vulnEntry.getStartLine()));
 
+        System.exit(2);
+
         // == Transform code / repair ==
         Map<String, List<JSONObject>> problemFixMap = new HashMap<>();
