+++ src/main/java/eu/assuremoss/framework/modules/analyzer/OpenStaticAnalyzer.java
@@ -47,6 +47,15 @@ public class OpenStaticAnalyzer implements CodeAnalyzer, VulnerabilityDetector, 
     private final String projectName;
     private final Map<String, String> supportedProblemTypes;
 
+    public static final HashMap<String, String> vulnMap = new HashMap<>() {{
+        // TODO: read from config file
+        put("FB_EiER", "EI_EXPOSE_REP2");
+        put("FB_EER", "EI_EXPOSE_REP2");
+        put("FB_NNPD", "NP_NULL_PARAM_DEREF");
+        put("FB_NNOSP", "NP_NULL_ON_SOME_PATH");
+        put("FB_MSBF", "MS_SHOULD_BE_FINAL");
+    }};
+
     @Override
     public List<CodeModel> analyzeSourceCode(File srcLocation, boolean isValidation) {
         PatchCompiler patchCompiler = PatchCompilerFactory.getPatchCompiler(VulnRepairDriver.properties.getProperty(PROJECT_BUILD_TOOL_KEY));
@@ -122,9 +131,6 @@ public class OpenStaticAnalyzer implements CodeAnalyzer, VulnerabilityDetector, 
 
     @Override
     public List<VulnerabilityEntry> getVulnerabilityLocations(File srcLocation, List<CodeModel> analysisResults) {
-        for (CodeModel cm : analysisResults) {
-            MLOG.info("Location: " + cm.getModelPath());
-        }
         List<VulnerabilityEntry> resList = new ArrayList<>();
         Optional<CodeModel> graphXML = analysisResults.stream()
                 .filter(cm -> cm.getType() == CodeModel.MODEL_TYPES.OSA_GRAPH_XML).findFirst();
@@ -152,38 +158,107 @@ public class OpenStaticAnalyzer implements CodeAnalyzer, VulnerabilityDetector, 
             List<Node> attributes = nodeListToArrayList(nodeList);
 
             for (Node node : attributes) {
-                String nodeName = getNodeAttribute(node, "name");
-                String context = getNodeAttribute(node, "context");
-                String problemType = supportedProblemTypes.get(nodeName);
+                if (!context(node).equals("warning") || problemType(node) == null) continue;
 
-                if (!context.equals("warning")) continue;
-                if (problemType == null) continue;
+                System.out.println("\n== Vulnerability: " + problemType(node) + " ==");
+                System.out.println("+ In file: " + filePath(node));
+                System.out.println("+ In line: " + lineNumStr(node));
+                System.out.println("+ Variable: " + variableName(findBugsXML, node));
 
-                MLOG.info("Found " + problemType + ", now trying to map column info...");
-
-                // TODO: replace .item(3) with actually finding the LineNum Node
-                String filePath = getNodeAttribute(node.getChildNodes().item(1), "value");
-                String lineNumStr = getNodeAttribute(node.getChildNodes().item(3), "value");
-
-                Pair<Integer, Integer> columnInfo = ColumnInfoParser
-                        .getColumnInfoFromFindBugsXML(filePath, nodeName, lineNumStr, findBugsXML.get());
-                MLOG.info("Line: " + lineNumStr + ", Column: [" + columnInfo.getA() + ", " + columnInfo.getB() + "]");
+                Pair<Integer, Integer> columnInfo = ColumnInfoParser.getColumnInfoFromFindBugsXML(filePath(node), nodeName(node), lineNumStr(node), variableName(findBugsXML, node));
 
                 NodeList warnAttributes = node.getChildNodes();
-                resList.add(createVulnerabilityEntry(warnAttributes, problemType, columnInfo));
+                resList.add(createVulnerabilityEntry(warnAttributes, problemType(node), columnInfo));
             }
-        } catch (FileNotFoundException e) {
-            LOG.error(e);
-        } catch (IOException e) {
-            LOG.error(e);
-        } catch (ParserConfigurationException e) {
-            LOG.error(e);
-        } catch (SAXException e) {
+        } catch (IOException | ParserConfigurationException | SAXException e) {
             LOG.error(e);
         }
 
         return resList;
     }
+
+    private String variableName(Optional<CodeModel> findBugsXML, Node node) {
+        return findVariableInFindBugsXML(nodeName(node), lineNumStr(node), findBugsXML.get());
+    }
+
+    public static String findVariableInFindBugsXML(String vulnType, String lineNum, CodeModel findBugsCM) {
+        if (findBugsCM.getType() != CodeModel.MODEL_TYPES.FINDBUGS_XML) return null;
+
+        DocumentBuilderFactory dbf = DocumentBuilderFactory.newInstance();
+        DocumentBuilder db;
+        Document doc;
+
+        try {
+            // process XML securely, avoid attacks like XML External Entities (XXE)
+            dbf.setFeature(XMLConstants.FEATURE_SECURE_PROCESSING, true);
+            db = dbf.newDocumentBuilder();
+            doc = db.parse(findBugsCM.getModelPath().getAbsolutePath());
+        } catch (Exception e) {
+            e.printStackTrace();
+            return null;
+        }
+
+        // TODO: clean this up
+        List<Node> bugInstances = nodeListToArrayList(doc.getElementsByTagName("BugInstance"));
+        for (Node bugInstance : bugInstances) {
+            String bugType = getNodeAttribute(bugInstance, "type");
+            if (!vulnMap.containsKey(vulnType)) continue; // vuln type is unsupported
+            if (!vulnMap.get(vulnType).equals(bugType)) continue;
+
+            String foundLineNum = null;
+            Node localVariable = null;
+
+            List<Node> children = nodeListToArrayList(bugInstance.getChildNodes());
+            for (Node child : children) {
+                // Get line num
+                switch (child.getNodeName()) {
+                    case "SourceLine":
+                        foundLineNum = getNodeAttribute(child, "start");
+                        break;
+
+                    case "LocalVariable":
+                    case "Field":
+                        localVariable = child;
+                        break;
+                }
+            }
+
+            // TODO: clean this up
+            if (foundLineNum != null && localVariable == null) {
+//                System.out.println("Found " + bugType + " on line " + foundLineNum + " without associated variable!");
+                return null;
+            }
+
+            if (foundLineNum == null) continue;
+
+            if (foundLineNum.equals(lineNum)) {
+                return getNodeAttribute(localVariable, "name");
+            }
+        }
+
+        return null;
+    }
+
+    private String lineNumStr(Node node) {
+        return getNodeAttribute(node.getChildNodes().item(3), "value");
+    }
+
+    private String filePath(Node node) {
+        return getNodeAttribute(node.getChildNodes().item(1), "value");
+    }
+
+    private String problemType(Node node) {
+        return supportedProblemTypes.get(nodeName(node));
+    }
+
+    private String context(Node node) {
+        return getNodeAttribute(node, "context");
+    }
+
+    private String nodeName(Node node) {
+        return getNodeAttribute(node, "name");
+    }
+
 
     private VulnerabilityEntry createVulnerabilityEntry(NodeList warnAttributes, String problemType, Pair<Integer, Integer> columnInfo) {
         VulnerabilityEntry ve = new VulnerabilityEntry();
